% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiread.R
\name{convert_df}
\alias{convert_df}
\title{convert_df}
\usage{
convert_df(df_list)
}
\arguments{
\item{df_list}{A list of dataframes}
}
\value{
A dataframe
}
\description{
convert_df
}
\examples{
\dontrun{
mydf <- convert_df(files)
}
}
