% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiread.R
\name{multiread}
\alias{multiread}
\title{multiread}
\usage{
multiread(filelist, id = TRUE, as_df = FALSE)
}
\arguments{
\item{filelist}{A character vector of csv files}

\item{id}{Logical, whether an id column should be included in the
dataframe}

\item{as_df}{Logical, whether to return a dataframe instead of a list
of dataframes}
}
\value{
A list with elements equal to `length(filelist)``
}
\description{
multiread
}
\examples{
\dontrun{
files <- list.files(pattern = ".csv")
mydata <- multiread(files, id = TRUE)
}
}
